+++ crates/filter-list-manager/src/utils/iterators/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod lines_with_terminator;